moved autoaccept to misc.cpp
