+++ DragonBurn/DragonBurn.vcxproj
@@ -180,7 +180,6 @@
     <ClInclude Include="Core\Init.h" />
     <ClInclude Include="Core\Render.h" />
     <ClInclude Include="Features\Aimbot.h" />
-    <ClInclude Include="Features\AutoAccept.h" />
     <ClInclude Include="Features\BombTimer.h" />
     <ClInclude Include="Features\ESP.h" />
     <ClInclude Include="Features\Misc.h" />
@@ -232,7 +231,6 @@
     <ClCompile Include="Core\Cheats.cpp" />
     <ClCompile Include="Core\GlobalVars.cpp" />
     <ClCompile Include="Features\Aimbot.cpp" />
-    <ClCompile Include="Features\AutoAccept.cpp" />
     <ClCompile Include="Features\Misc.cpp" />
     <ClCompile Include="Features\Radar.cpp" />
     <ClCompile Include="Features\RCS.cpp" />
